requirements.txt
psycopg>=3.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
metaflow>=2.8.0
pandas>=2.0.0
loguru>=0.7.0